Move SAI services to sub-folders
