+++ pod-provider/backend/services/sai/auth-registry/auth-registry.js
